refactor use cases
