+++ service_taxi/src/main/kotlin/org/thechance/service_taxi/api/endpoints/TaxiRoutes.kt
@@ -15,44 +15,42 @@ import org.thechance.service_taxi.api.models.taxi.TaxiDto
 import org.thechance.service_taxi.api.models.taxi.toDto
 import org.thechance.service_taxi.api.models.taxi.toEntity
 import org.thechance.service_taxi.api.models.taxi.toUpdateRequest
-import org.thechance.service_taxi.domain.usecase.AdministratorUseCase
-import org.thechance.service_taxi.domain.usecase.CustomerUseCase
+import org.thechance.service_taxi.domain.usecase.IManageTaxiUseCase
 import org.thechance.service_taxi.domain.util.MissingParameterException
 
 fun Route.taxiRoutes() {
-    val administratorUseCase: AdministratorUseCase by inject()
-    val customerUseCase: CustomerUseCase by inject()
+    val manageTaxiUseCase: IManageTaxiUseCase by inject()
 
     route("/taxi") {
         get {
             val page = call.parameters["page"]?.toInt() ?: 1
             val limit = call.parameters["limit"]?.toInt() ?: 20
-            val result = customerUseCase.getAllTaxi(page, limit)
+            val result = manageTaxiUseCase.getAllTaxi(page, limit)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
         get("/{taxiId}") {
             val id = call.parameters["taxiId"] ?: throw MissingParameterException
-            val result = customerUseCase.getTaxi(id)
+            val result = manageTaxiUseCase.getTaxi(id)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
         post {
             val taxi = call.receive<TaxiDto>()
-            administratorUseCase.createTaxi(taxi.toEntity())
+            manageTaxiUseCase.createTaxi(taxi.toEntity())
             call.respond(HttpStatusCode.OK, "added")
         }
 
         put("/{taxiId}") {
             val taxiId = call.parameters["taxiId"] ?: throw MissingParameterException
             val taxi = call.receive<TaxiDto>()
-            administratorUseCase.updateTaxi(taxi.toUpdateRequest().copy(id = taxiId))
+            manageTaxiUseCase.updateTaxi(taxi.toUpdateRequest().copy(id = taxiId))
             call.respond(HttpStatusCode.OK, "updated")
         }
 
         delete("/{taxiId}") {
             val taxiId = call.parameters["taxiId"] ?: throw MissingParameterException
-            administratorUseCase.deleteTaxi(taxiId)
+            manageTaxiUseCase.deleteTaxi(taxiId)
             call.respond(HttpStatusCode.OK, "deleted")
         }
     }
