+++ service_taxi/src/main/kotlin/org/thechance/service_taxi/api/endpoints/TripRoutes.kt
@@ -14,28 +14,26 @@ import org.koin.ktor.ext.inject
 import org.thechance.service_taxi.api.models.trip.TripDto
 import org.thechance.service_taxi.api.models.trip.toDto
 import org.thechance.service_taxi.api.models.trip.toEntity
-import org.thechance.service_taxi.domain.usecase.AdministratorUseCase
-import org.thechance.service_taxi.domain.usecase.CustomerUseCase
-import org.thechance.service_taxi.domain.usecase.DriverUseCase
+import org.thechance.service_taxi.domain.usecase.IDiscoverTripsUseCase
+import org.thechance.service_taxi.domain.usecase.IManageTripsUseCase
 import org.thechance.service_taxi.domain.util.MissingParameterException
 
 fun Route.tripRoutes() {
-    val driverUseCase: DriverUseCase by inject()
-    val administratorUseCase: AdministratorUseCase by inject()
-    val customerUseCase: CustomerUseCase by inject()
+    val manageTripsUseCase: IManageTripsUseCase by inject()
+    val discoverTripsUseCase: IDiscoverTripsUseCase by inject()
 
 
     route("/trip") {
         get {
             val page = call.parameters["page"]?.toInt() ?: 1
             val limit = call.parameters["limit"]?.toInt() ?: 20
-            val result = administratorUseCase.getTrips(page, limit)
+            val result = manageTripsUseCase.getTrips(page, limit)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
         get("/{tripId}") {
             val id = call.parameters["tripId"] ?: throw MissingParameterException
-            val result = customerUseCase.getTripById(id)
+            val result = manageTripsUseCase.getTripById(id)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
@@ -43,7 +41,7 @@ fun Route.tripRoutes() {
             val id = call.parameters["driverId"] ?: throw MissingParameterException
             val page = call.parameters["page"]?.toInt() ?: 1
             val limit = call.parameters["limit"]?.toInt() ?: 20
-            val result = driverUseCase.getTripsByDriverId(id, page, limit)
+            val result = discoverTripsUseCase.getTripsByDriverId(id, page, limit)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
@@ -51,20 +49,20 @@ fun Route.tripRoutes() {
             val id = call.parameters["clientId"] ?: throw MissingParameterException
             val page = call.parameters["page"]?.toInt() ?: 1
             val limit = call.parameters["limit"]?.toInt() ?: 20
-            val result = customerUseCase.getTripsByClientId(id, page, limit)
+            val result = discoverTripsUseCase.getTripsByClientId(id, page, limit)
             call.respond(HttpStatusCode.OK, result.toDto())
         }
 
         post {
             val tripDto = call.receive<TripDto>()
-            customerUseCase.createTrip(tripDto.toEntity())
+            manageTripsUseCase.createTrip(tripDto.toEntity())
             call.respond(HttpStatusCode.Created, "added")
         }
 
         put("/{tripId}/rate") {
             val tripId = call.parameters["tripId"] ?: throw MissingParameterException
             val rate = call.parameters["rate"]?.toDouble() ?: throw MissingParameterException
-            customerUseCase.rateTrip(tripId, rate)
+            manageTripsUseCase.rateTrip(tripId, rate)
             call.respond(HttpStatusCode.OK, "updated")
         }
 
@@ -72,20 +70,20 @@ fun Route.tripRoutes() {
             val tripId = call.parameters["tripId"] ?: throw MissingParameterException
             val driverId = call.parameters["driverId"] ?: throw MissingParameterException
             val taxiId = call.parameters["taxiId"] ?: throw MissingParameterException
-            driverUseCase.approveTrip(driverId, taxiId, tripId)
+            discoverTripsUseCase.approveTrip(driverId, taxiId, tripId)
             call.respond(HttpStatusCode.OK, "updated")
         }
 
         put("/{tripId}/finish") {
             val tripId = call.parameters["tripId"] ?: throw MissingParameterException
             val driverId = call.parameters["driverId"] ?: throw MissingParameterException
-            driverUseCase.finishTrip(driverId, tripId)
+            discoverTripsUseCase.finishTrip(driverId, tripId)
             call.respond(HttpStatusCode.OK, "updated")
         }
 
         delete("/{tripId}") {
             val tripId = call.parameters["tripId"] ?: throw MissingParameterException
-            administratorUseCase.deleteTrip(tripId)
+            manageTripsUseCase.deleteTrip(tripId)
             call.respond(HttpStatusCode.OK, "deleted")
         }
     }
