+++ service_taxi/src/main/kotlin/org/thechance/service_taxi/domain/usecase/DiscoverTripsUseCase.kt
@@ -9,10 +9,17 @@ import org.thechance.service_taxi.domain.entity.TripUpdateRequest
 import org.thechance.service_taxi.domain.gateway.DataBaseGateway
 import org.thechance.service_taxi.domain.util.ResourceNotFoundException
 
+interface IDiscoverTripsUseCase {
+    suspend fun getTripsByDriverId(driverId: String, page: Int, limit: Int): List<Trip>
+    suspend fun getTripsByClientId(clientId: String, page: Int, limit: Int): List<Trip>
+    suspend fun approveTrip(driverId: String, taxiId: String, tripId: String)
+    suspend fun finishTrip(driverId: String, tripId: String)
+}
+
 @Single
-class DriverUseCaseImp(
-    private val dataBaseGateway: DataBaseGateway,
-) : DriverUseCase {
+class DiscoverTripsUseCase(
+    private val dataBaseGateway: DataBaseGateway
+) : IDiscoverTripsUseCase {
     override suspend fun approveTrip(driverId: String, taxiId: String, tripId: String) {
         dataBaseGateway.getTripById(tripId) ?: throw ResourceNotFoundException
         dataBaseGateway.updateTrip(
@@ -34,12 +41,18 @@ class DriverUseCaseImp(
             )
         )
     }
-
     override suspend fun getTripsByDriverId(
         driverId: String,
         page: Int,
         limit: Int
     ): List<Trip> {
         return dataBaseGateway.getDriverTripsHistory(driverId, page, limit)
+    }
+    override suspend fun getTripsByClientId(
+        clientId: String,
+        page: Int,
+        limit: Int
+    ): List<Trip> {
+        return dataBaseGateway.getClientTripsHistory(clientId, page, limit)
     }
 }
